WatcherWindow turn into event driven.
I check the code in WatcherWindow.java, it's using JavaFX in a very
very very bad way. If you didn't understand it properly please do not
use this technology. I think the JavaFX offered an observed object that
allows you to add a listener on the value change event. However, the
code in WatcherWindow didn't properly follow the document from the
Oracle
http://docs.oracle.com/javafx/2/collections/jfxpub-collections.htm In
contrast, it used the solution came from
http://stackoverflow.com/questions/11065140/javafx-2-1-tableview-refresh
-items . Remove all the element? And set invisible and visible agian?
Please, don't do that, there is a method for Tableview name refresh, it
will refresh all the table using new data. Back to the problem itself.
Simply do the tableview refresh instead of removing all the element and
readd agian and invisible visible can solve the problem. However, it
remains ugly, not an event driven method. We have already used the
Oberservable Object, why not be totally event driven? Sure we can,
according to the official manual, the underlying object won't be
observered, that is the value inside the this won't be observed. So we
need to add a Listener to the value itself. That's it, fixed the
problem.