Decouple SwitchGUI and SwitchDevice
With the help of Google EvenBus, SwitchGUI is now independent of
SwitchDevice.
1, Add: DeviceInputEvent.java, Introducing one new Event type,
DeviceInputEvent, in the construction function you need to specify the
deviceName and the deviceData passing to that device.
2, Modify: Read the deviceName from SetupDevices (Todo, read from
configuration and generate by address bus in the future)
3, Add: Switches.java each device should contain their own name as id
(Todo, id assigned by address bus, while attach the device into the bus
to allow multiple same device). The device subscribe to the
inputFromGUI channel, and based on the deviceName to distinguish from
different source. At last, start listening on EventBus.
4, Modify: EmulationWindow.java passing deviceName to the constructor
(Todo, read from configuration and generate by address bus in the
future)
5, Modify: SwitchesDisplay.java removed the switchBackend from the
Display(GUI) part.