Decoupled the ASMFile from GUI Editor
1. UpdateASMEvent.java Add one more event: UpdateASMEvent indicate that
there is a code updating in GUI Editor
2. CodeEditor.java Accept projectName fileName and fileContent, the
newly created CodeEditor will emit an event distinguished by
projectName and fileName when the Text inside the Editor was updated.
3. Main.java, change the logic of creating a new CodeEditor, passing
the content and projectName, fileName to the CodeEditor. Remove the
bind content to ASMFile operation, since the CodeEditor will emit an
event, and the ASMFile will listen on that event to receive updated
fileContent.
4. SimpleASMFile.java when creating a new file, listening on the update
event on specified file(Unique by projectName and fileName).